+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @palantir/pack.schema
 
+## 0.6.0
+
+### Minor Changes
+
+- b114b1a: Propagate lenses explicitly
+
+## 0.5.0
+
+### Minor Changes
+
+- 3c7a9f3: Split IR Versioned Tasks
+- 1ebb677: Always use the IR as intermediary between builders and SDK/Zod/Wire
+
 ## 0.4.0
 
 ### Minor Changes